requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-ollama
langchain_google_genai
python-dotenv
pydantic
duckduckgo-search